refactor(desktop): move the task composer row into its own file
TasksPage.swift is under a line-count ratchet, and the composer is a
self-contained view that the page, the Today section, and the row-anchored
insert all render. It belongs beside the other Tasks views rather than in
the page file.
